vpn: do not remove .info file on disconnect
The info file contains the name of the unit along with the installed
version.
Such information are useful also when the unit has been temporary
disconnected from the controller.

If the file is not present, the user can't see the name of the
disconnected user but just the uuid which can be hardly linked
to the unit name.