feat(api)!: move JWT tokens to RAM
This change will prevent that someone can access valid tokens
looking inside the filesytem.

On NS8, the module did not use a volume for the token directory,
so the behavior does not change: on controller restart, all sessions
are destroyed.

This change is part of a refactor to improve security.